Fixed the signInUser mutation
SCHEMA/RESOLVERS.JS:
    ln 68: findOne was raising an error indicating that options.where was
missing. fixed to now properly select a user where email matches

    ln 69: user.checkPassword custom method returns a promise. it needed to
be awaited to properly return the expected boolean

MODELS/USER.JS:
    ln 90: checkPassword was raising an error signifying that invalid
inputs were being passed to the bcrypt.compare() method. it was
displayed as STRING, UNDEFINED indicating that the 'this' keyword was
failing within the instance method. modified the checkPassword method to
accept a current_user parameter which is passed as the user instance
generated in the resolver. this provides the expected 'this'
functionality - no pun intended...